requirements.txt
genshin
python-telegram-bot==20.0a0
sentry-sdk
asyncio